- Applied a GCC attribute to methods that take format strings (CFileText::Printf, CEventLog::Event*, CTextConsole::SysMessagef, etc), which enables the compiler to validate the format strings and their parameters. - Addressed a number of GCC warnings, mostly regarding the improper use of %d/%i/%x/%u for long int types (should be using %ld/%li/%lx/%lu) and several cases where a format method was being using without any arguments. Nothing has changed functionally. Not updated any of the format strings contained in defmessages for now to avoid having to update sphere_msgs.scp. - Added a cast in mtrand.h to fix a GCC warning - would prefer to avoid modifying 3rd party files but GCC insists on throwing it out on every single cpp file. - Updated linux + bsd makefiles.

+++ 0.56.b/common/mtrand/mtrand.h
@@ -215,7 +215,7 @@ inline void MTRand::seed( uint32 *const bigSeed, const uint32 seedLength )
 	initialize(19650218UL);
 	register int i = 1;
 	register uint32 j = 0;
-	register int k = ( N > seedLength ? N : seedLength );
+	register int k = ( N > seedLength ? static_cast<uint32>(N) : seedLength );
 	for( ; k; --k )
 	{
 		state[i] =
